+++ babel.config.js
@@ -3,13 +3,14 @@
 // actually result in valid code for Node to run.
 module.exports = {
   plugins: [
+    '@babel/plugin-proposal-numeric-separator',
     '@babel/plugin-proposal-optional-chaining',
     '@babel/plugin-proposal-nullish-coalescing-operator',
-    ['@babel/plugin-proposal-class-properties', {loose: true}],
-    '@babel/plugin-proposal-numeric-separator',
+    '@babel/plugin-proposal-class-properties',
   ],
+  targets: 'current node',
   presets: [
-    ['@babel/preset-env', {targets: {node: 'current'}, modules: 'commonjs'}],
+    ['@babel/preset-env', {modules: 'commonjs'}],
     '@babel/preset-typescript',
   ],
 };