+++ packages/plugin-typescript/CHANGELOG.md
@@ -10,6 +10,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Update `@babel/*` packages to their latest versions `7.14.5` [[#191](https://github.com/Shopify/sewing-kit-next/pull/191)]
+- Remove `babel-plugin-convert-empty-file-to-esmodule`. This feature is now provided by `@babel/plugin-transform-typescript` [[#195](https://github.com/Shopify/sewing-kit-next/pull/195)]
 
 ## 0.2.1 - 2021-05-28