+++ packages/plugin-typescript/CHANGELOG.md
@@ -10,7 +10,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Added package metadata and limit files included in packages. Ensure `*.tsbuildinfo` files are absent to greatly shink package size. [[#177](https://github.com/Shopify/sewing-kit-next/pull/177)]
-- Removed configuration of eslint extensions, which is no longer required as typescript files can be included in linting by modifying ESLint config files.
+- Removed configuration of eslint extensions, which is no longer required as typescript files can be included in linting by modifying ESLint config files. [[#199](https://github.com/Shopify/sewing-kit-next/pull/199)]
 
 ## 0.3.0 - 2021-06-18